Create an abstract base class for drawable catch objects
